cli structure refactor
