+++ src/cli/isCommentPresent.ts
@@ -1,0 +1,15 @@
+import { readFileSync } from 'fs';
+
+export function isCommentPresent(commentText: string, filePath: string): boolean {
+  const allLines = readFileSync(filePath).toString().split('\n');
+  const comments = allLines.filter((line) => line.startsWith('//'));
+
+  return comments.some((comment) =>
+    Array.from(comment)
+      .filter((char) => char !== '/')
+      .join('')
+      .trim()
+      .split(' ')
+      .includes(commentText),
+  );
+}
