+++ src/cli/tsc-strict/findStrictFiles.ts
@@ -1,6 +1,7 @@
-import { getPosixFilePath } from '../common/utils';
-import * as typescript from './typescript/typescript';
-import { CliStrictFileChecker, getPluginConfig } from './CliStrictFileChecker';
+import { getPosixFilePath } from '../../common/utils';
+import * as typescript from '../typescript/typescript';
+import { CliStrictFileChecker } from './CliStrictFileChecker';
+import { getPluginConfig } from '../getPluginConfig';
 
 export async function findStrictFiles(): Promise<string[]> {
   const filesCheckedByTS = await getFilesCheckedByTs();
