+++ src/common/getAbsolutePath.ts
@@ -1,0 +1,6 @@
+import { isAbsolute, resolve } from 'path';
+
+export function getAbsolutePath(projectRootPath: string, filePath: string): string {
+  if (isAbsolute(filePath)) return filePath;
+  return resolve(projectRootPath, filePath);
+}
