+++ src/common/isFileOnPath.ts
@@ -1,5 +1,5 @@
 import { getPosixFilePath } from './utils';
-import { getAbsolutePath } from '../cli/CliStrictFileChecker';
+import { getAbsolutePath } from './getAbsolutePath';
 
 export function isFileOnPath(currentFilePath: string, pathToStrictFiles: string): boolean {
   const absolutePathToStrictFiles = getAbsolutePath(process.cwd(), pathToStrictFiles);
